cdk/requirements.txt
aws-cdk-lib
cdk-ecr-deployment
boto3
